# backend/openapi/UKPS.Api.json
{
  "openapi": "3.1.1",
  "info": {
    "title": "UKPS.Api | v1",
    "version": "1.0.0"
  },
  "paths": {
...
    "/users/me/current-organisation": {
      "patch": {
        "tags": [
          "User"
        ],
        "summary": "Updates the current user's selected organisation.",
        "operationId": "UpdateCurrentOrganisation",
        "requestBody": {
          "description": "A token that can be used to cancel the request.",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateCurrentOrganisationCommand"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateCurrentOrganisationCommand"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateCurrentOrganisationCommand"
              }
            }
          },
          "required": true
        },
        "responses": {
          "400": {
            "description": "The provided organisation is invalid.",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "200": {
            "description": "The current organisation was successfully updated."
          }
        }
      }
    },
    "/users": {
      "get": {
        "tags": [
          "User"
        ],
        "summary": "Retrieves a paginated list of users based on the specified query parameters.",
        "operationId": "GetUsers",
        "parameters": [
          {
            "name": "OrganisationId",
            "in": "query",
            "description": "Gets or initialises the optional ID of the organisation to filter users by.",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "Page",
            "in": "query",
            "description": "Gets or initialises the page number for pagination.\nMust be greater than or equal to 1.",
            "schema": {
              "maximum": 2147483647,
              "minimum": 1,
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "PageSize",
            "in": "query",
            "description": "Gets or initialises the number of items per page for pagination.\nMust be between 1 and 100.",
            "schema": {
              "maximum": 100,
              "minimum": 1,
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "Status",
            "in": "query",
            "description": "Gets or initialises the collection of user organisation statuses to filter users by.\nUsers with a UserOrgStatus.Rejected status are always excluded from\nresults, regardless of the values supplied here.",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/components/schemas/UserOrgStatus"
              }
            }
          },
          {
            "name": "Role",
            "in": "query",
            "description": "Gets or initialises the collection of user roles to filter users by.",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/components/schemas/UserRole"
              }
            }
          },
          {
            "name": "Email",
            "in": "query",
            "description": "Gets or initialises an optional partial, case-insensitive match against the user's email address.",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "LastActiveFrom",
            "in": "query",
            "description": "Gets or initialises the inclusive lower bound of the user's last active timestamp to filter by.",
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          },
          {
            "name": "LastActiveTo",
            "in": "query",
            "description": "Gets or initialises the inclusive upper bound of the user's last active timestamp to filter by.",
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          },
          {
            "name": "SortBy",
            "in": "query",
            "description": "Gets or initialises the field by which users are sorted.\nDefaults to GetUsersQuerySortValue.LastActive.",
            "schema": {
              "$ref": "#/components/schemas/GetUsersQuerySortValue"
            }
          },
          {
            "name": "SortDirection",
            "in": "query",
            "description": "Gets or initialises the direction in which users are sorted.\nDefaults to SortDirection.Descending.",
            "schema": {
              "$ref": "#/components/schemas/SortDirection"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Returns the paginated list of users.",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedResponseDtoOfUserListItemDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedResponseDtoOfUserListItemDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedResponseDtoOfUserListItemDto"
                }
              }
            }
          },
          "400": {
            "description": "Returned if the query parameters are invalid or the organisation is not found.",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "Returned if no users are found matching the query parameters.",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/users/{userId}/organisations/{organisationId}": {
      "get": {
        "tags": [
          "User"
        ],
        "summary": "Retrieves a user's details along with their role within the specified organisation.",
        "operationId": "GetUserDetailsWithinOrganisation",
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "description": "The unique identifier of the user to retrieve.",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "organisationId",
            "in": "path",
            "description": "The unique identifier of the organisation to read the user's membership from. A user may\nbelong to several organisations, so their role is resolved against this organisation.",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Returns the user's details and their role within the organisation.",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/UserInformationDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserInformationDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserInformationDto"
                }
              }
            }
          },
          "400": {
            "description": "Returned if the specified organisation does not exist.",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "403": {
            "description": "Returned if the caller is not authorised to view the organisation's users.",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "Returned if the user is not a member of the specified organisation.",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/users/{userId}": {
      "patch": {
        "tags": [
          "User"
        ],
        "summary": "Updates the details of the specified user.",
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "description": "The unique identifier of the user whose details are being updated.",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "description": "A token to monitor for cancellation requests.",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateUserDetailsCommand"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateUserDetailsCommand"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateUserDetailsCommand"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "The user's details were successfully updated.",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/UserDetailsDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserDetailsDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserDetailsDto"
                }
              }
            }
          },
          "400": {
            "description": "The supplied user details are invalid.",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "The specified user does not exist.",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "403": {
            "description": "The caller is not authorised to update the specified user's details.",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "409": {
            "description": "The request conflicts with the existing data such as another users email.",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "GetUsersQuerySortValue": {
        "enum": [
          "LastActive",
          "Email",
          "Role",
          "Status"
        ],
        "type": "string",
        "description": "Specifies the fields by which users can be sorted when querying users."
      },
...
      "PaginatedResponseDtoOfUserListItemDto": {
        "required": [
          "items",
          "totalCount",
          "page",
          "pageSize"
        ],
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/UserListItemDto"
            },
            "description": "Gets the collection of items in the current page."
          },
          "totalCount": {
            "type": "integer",
            "description": "Gets the total number of items across all pages.",
            "format": "int32"
          },
          "page": {
            "type": "integer",
            "description": "Gets the current page number (1-based index).",
            "format": "int32"
          },
          "pageSize": {
            "type": "integer",
            "description": "Gets the number of items per page.",
            "format": "int32"
          }
        },
        "description": "Represents a paginated response containing a collection of items and pagination metadata."
      },
...
      "ProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": [
              "null",
              "string"
            ]
          },
          "title": {
            "type": [
              "null",
              "string"
            ]
          },
          "status": {
            "type": [
              "null",
              "integer"
            ],
            "format": "int32"
          },
          "detail": {
            "type": [
              "null",
              "string"
            ]
          },
          "instance": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
...
      "SortDirection": {
        "enum": [
          "Ascending",
          "Descending"
        ],
        "type": "string",
        "description": "Specifies the direction in which query results are sorted."
      },
...
      "UpdateCurrentOrganisationCommand": {
        "required": [
          "organisationId"
        ],
        "type": "object",
        "properties": {
          "organisationId": {
            "type": "integer",
            "description": "Gets the identifier of the organisation to set as the current organisation.",
            "format": "int32"
          }
        },
        "description": "Represents the request to update the currently authenticated user's active organisation."
      },
...
      "UpdateUserDetailsCommand": {
        "required": [
          "fullName",
          "workEmail",
          "workTelephone"
        ],
        "type": "object",
        "properties": {
          "fullName": {
            "type": "string",
            "description": "Gets the user's full name."
          },
          "workEmail": {
            "type": "string",
            "description": "Gets the user's work email address."
          },
          "workTelephone": {
            "type": "string",
            "description": "Gets the user's work telephone number."
          }
        },
        "description": "Represents the details to update for an existing user."
      },
      "UserDetailsDto": {
        "required": [
          "userType",
          "title",
          "fullName",
          "jobTitle",
          "workPhone",
          "workEmail"
        ],
        "type": "object",
        "properties": {
          "userType": {
            "description": "Gets the user's type.",
            "$ref": "#/components/schemas/UserType"
          },
          "title": {
            "type": [
              "null",
              "string"
            ],
            "description": "Gets the user's title (for example, Mr, Mrs, Ms, or Dr), if available."
          },
          "fullName": {
            "type": "string",
            "description": "Gets the user's first name."
          },
          "jobTitle": {
            "type": [
              "null",
              "string"
            ],
            "description": "Gets the user's job title, if available."
          },
          "workPhone": {
            "type": [
              "null",
              "string"
            ],
            "description": "Gets the user's work telephone number, if available."
          },
          "workEmail": {
            "type": "string",
            "description": "Gets the user's work email address."
          }
        },
        "description": "Represents the details of a user."
      },
      "UserInformationDto": {
        "required": [
          "userId",
          "fullName",
          "workTelephone",
          "workEmail",
          "organisationMembershipId",
          "organisationId",
          "organisationName",
          "userRole"
        ],
        "type": "object",
        "properties": {
          "userId": {
            "type": "integer",
            "description": "Gets the unique identifier of the user.",
            "format": "int32"
          },
          "fullName": {
            "type": "string",
            "description": "Gets the full name of the user."
          },
          "workTelephone": {
            "type": "string",
            "description": "Gets the user's work telephone number."
          },
          "workEmail": {
            "type": "string",
            "description": "Gets the user's email address."
          },
          "organisationMembershipId": {
            "type": "integer",
            "description": "Gets the unique identifier of the user's organisation membership.",
            "format": "int32"
          },
          "organisationId": {
            "type": "integer",
            "description": "Gets the unique identifier of the user's organisation.",
            "format": "int32"
          },
          "organisationName": {
            "type": "string",
            "description": "Gets the name of the user's organisation."
          },
          "userRole": {
            "description": "Gets the role assigned to the user within the organisation.",
            "$ref": "#/components/schemas/UserRole"
          }
        },
        "description": "Represents information about a user and their membership within an organisation."
      },
      "UserListItemDto": {
        "required": [
          "userId",
          "registrationRequestId",
          "emailAddress",
          "role",
          "status",
          "lastActive",
          "actions"
        ],
        "type": "object",
        "properties": {
          "userId": {
            "type": [
              "null",
              "integer"
            ],
            "description": "Gets the unique identifier of the user.",
            "format": "int32"
          },
          "registrationRequestId": {
            "type": [
              "null",
              "integer"
            ],
            "description": "The optional lasted active registration request associated with the user.",
            "format": "int32"
          },
          "emailAddress": {
            "type": "string",
            "description": "Gets the email address of the user, if available."
          },
          "role": {
            "description": "Gets the role of the user within the system.",
            "$ref": "#/components/schemas/UserRole"
          },
          "status": {
            "description": "Gets the organisational status of the user.",
            "$ref": "#/components/schemas/UserOrgStatus"
          },
          "lastActive": {
            "type": [
              "null",
              "string"
            ],
            "description": "Gets the date and time when the user was last active, if available.",
            "format": "date-time"
          },
          "actions": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/UserMembershipAction"
            },
            "description": "Gets the actions that can be performed by the current user."
          }
        },
        "description": "Represents a limited summary of a user."
      },
...
      "UserOrgStatus": {
        "enum": [
          "RequestedAccess",
          "AwaitingSetup",
          "Active",
          "Rejected",
          "Inactive",
          "Deactivated"
        ],
        "type": "string",
        "description": "Represents the status of a user's association with an organisation."
      },
      "UserRole": {
        "enum": [
          "Standard",
          "Champion",
          "Super"
        ],
        "type": "string",
        "description": "Represents the different roles a user can have within the system."
      },
...
      "ValidationProblemDetails": {
        "required": [
          "errors"
        ],
        "type": "object",
        "properties": {
          "type": {
            "type": [
              "null",
              "string"
            ]
          },
          "title": {
            "type": [
              "null",
              "string"
            ]
          },
          "status": {
            "type": [
              "null",
              "integer"
            ],
            "format": "int32"
          },
          "detail": {
            "type": [
              "null",
              "string"
            ]
          },
          "instance": {
            "type": [
              "null",
              "string"
            ]
          },
          "errors": {
            "type": "object",
            "additionalProperties": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          }
        }
      },
...
    }
  },
...
}